sched/walt: Remove pipeline heavy tasks under heavy_lock
When a task dies, we check if its low_latency flag is set to
WALT_LOW_LATENCY_HEAVY_BIT or not. If it is, we remove the task from the
heavy_wts array by acquiring the heavy_lock.

However, there is a potential race condition where a task might die
after being added to heavy_wts but before its low_latency flag is set to
WALT_LOW_LATENCY_HEAVY_BIT. In such cases, the task will not be removed
from heavy_wts in the task termination code path. Thus, the task remains
in WALT heavy_wts list and may lead to access of task structure which is
already freed (use after free).

To fix this issue, move the check for WALT_LOW_LATENCY_HEAVY_BIT inside
the heavy_lock in the task termination code path. This ensures that a
dying pipeline low latency task is always removed from heavy_wts.

Change-Id: I9523ca80bdfaa52b1329ebee59a5ff39d8dc44d3
Signed-off-by: Atul Pant <quic_atulpant@quicinc.com>