sched/walt: Leverage zone based mechanism for cpufreq
By default WALT governor hops to next freq level once load reaches 80%
of the max load supported by current freq. This 80% threshold is the
default value and is common for all zones.

With the introduction of zone based mechanism, the target load
percentage values may vary and appropriately used for calculations of
the next frequency for a cluster.

Use zone based sysfs tuning for mapping util to freq. Also, Use
80% threshold as fallback for unspecified zones.

Change-Id: I210c864e8de238da7dbab685a6b7c283f94b6b9c
Signed-off-by: Ankit Sharma <quic_anshar@quicinc.com>