rpmsg: glink: Clear channel buffer for intentless use case
For the intentless channel use case, if the channel buffer
is not cleared after processing a complete RX message,
it may lead to side effects of handling new RX data
in the previous intent buffer.

Set channel->buf to NULL in case of an intentless
use case after processing the RX message.

Change-Id: I70bcd2c89b846c882f42139f7235a5ad8774db5d
Signed-off-by: Sudeepgoud Patil <quic_sudeepgo@quicinc.com>