ufs: ufs-qcom: Restrict esi-mask to valid CPUs
This update ensures that the esi-mask uses a subset of
cpu_possible_mask when the provided mask is invalid,
thereby restricting the ESI-mask to include only valid CPUs.

Previously, invalid CPU bits in the esi-mask could cause
memory aborts and kernel paging request errors if any CPU
in the esi-mask was not populated or was intentionally
removed.

Change-Id: I1e22339bdd0b6bebb23885fcba8527f0d0c17d3c
Signed-off-by: Manish Pandey <quic_mapa@quicinc.com>